Extract client logics to library
