+++ src/Kms.Client.Dispatcher/Kms.Client.Dispatcher.csproj
@@ -1,12 +1,11 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+    <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.3" />
-    <PackageReference Include="NLog" Version="4.7.2" />
     <PackageReference Include="NLog.Web.AspNetCore" Version="4.9.2" />
     <PackageReference Include="Google.Protobuf" Version="3.12.2" />
     <PackageReference Include="Grpc.Net.Client" Version="2.29.0" />
@@ -17,16 +16,11 @@
     </PackageReference>
     <PackageReference Include="System.Reactive" Version="4.4.1" />
   </ItemGroup>
-
+  
   <ItemGroup>
-    <ProjectReference Include="..\Kms.Client.Dispatcher\Kms.Client.Dispatcher.csproj" />
     <ProjectReference Include="..\Kms.Core\Kms.Core.csproj" />
     <ProjectReference Include="..\Kms.Crypto\Kms.Crypto.csproj" />
     <ProjectReference Include="..\Kms.KeyMngr\Kms.KeyMngr.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Protobuf Include="Protos\greet.proto" GrpcServices="Client" />
   </ItemGroup>
 
 </Project>
