+++ src/Kms.Client.Dispatcher/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Kms.Client.Dispatcher
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
