+++ src/Kms.Client.Lib/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Kms.Client.Lib
+{
+    public class Class1
+    {
+    }
+}
