+++ src/Kms.gRPC.sln
@@ -21,6 +21,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kms.gRPC", "Kms.gRPC\Kms.gR
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kms.gRPC.Client", "Kms.gRPC.Client\Kms.gRPC.Client.csproj", "{26A20098-4574-4028-8018-566CDAE8D7C5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kms.Client.Dispatcher", "Kms.Client.Dispatcher\Kms.Client.Dispatcher.csproj", "{E03D121D-F636-4BD1-804F-DB3A0DDE0B80}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,6 +49,10 @@ Global
 		{26A20098-4574-4028-8018-566CDAE8D7C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{26A20098-4574-4028-8018-566CDAE8D7C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{26A20098-4574-4028-8018-566CDAE8D7C5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E03D121D-F636-4BD1-804F-DB3A0DDE0B80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E03D121D-F636-4BD1-804F-DB3A0DDE0B80}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E03D121D-F636-4BD1-804F-DB3A0DDE0B80}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E03D121D-F636-4BD1-804F-DB3A0DDE0B80}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -57,6 +63,7 @@ Global
 		{60617002-D8B4-4455-BD42-8CBDC2631F4F} = {47765B9E-6940-4332-87F1-F47598586747}
 		{D4CB6E6C-3D02-4791-A43C-66421FE064CB} = {5F32E7B8-DD3F-42D7-A47E-93D6AB0B7E08}
 		{26A20098-4574-4028-8018-566CDAE8D7C5} = {5F32E7B8-DD3F-42D7-A47E-93D6AB0B7E08}
+		{E03D121D-F636-4BD1-804F-DB3A0DDE0B80} = {407374AC-6D0F-497B-A2FE-77075CE08809}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1DC8EF4A-588B-4CBA-AE76-18898F5F99F9}
